third_party/WebKit/LayoutTests/platform/win7/virtual/outofblink-cors-http/http/tests/fetch/serviceworker-proxied/thorough/redirect-nocors-other-https-expected.txt

This is a testharness.js-based test.
PASS Startup 
FAIL Verify access control of fetch() in a Service Worker assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Flocalhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=no-cors&method=GET&headers=CUSTOM jsonpResult must be success expected "success" but got "error"
Harness: the test ran to completion.
